chore (search-api) Separate DatasetsSearchController
