+++ applications/search-api/src/test/java/no/dcat/portal/query/controller/DatasetsSearchControllerTest.java
@@ -29,9 +29,9 @@ import static org.mockito.Mockito.*;
  */
 @RunWith(MockitoJUnitRunner.class)
 @Category(UnitTest.class)
-public class DatasetsQueryControllerSearchTest {
+public class DatasetsSearchControllerTest {
 
-    DatasetsQueryController sqs;
+    DatasetsSearchController sqs;
     Client client;
 
     @Before
@@ -40,7 +40,7 @@ public class DatasetsQueryControllerSearchTest {
         populateMock();
         ElasticsearchService elasticsearchServiceMock = mock(ElasticsearchService.class);
         when(elasticsearchServiceMock.getClient()).thenReturn(client);
-        sqs = new DatasetsQueryController(elasticsearchServiceMock);
+        sqs = new DatasetsSearchController(elasticsearchServiceMock);
     }
 
     /**
